Added weights

+++ ArrayFoo/app.py
@@ -1,7 +1,6 @@
 from flask import Flask, redirect, url_for, render_template, Response
 from model import *
 
-
 app = Flask(__name__)
 
 camera = cv.VideoCapture(0, cv.CAP_DSHOW)
@@ -20,7 +19,10 @@ def gen_frames():
             frame_resized = tf.image.resize_with_pad(
                 frame, config['INPUT_shape'][0], config['INPUT_shape'][0])[tf.newaxis, ...]
             pred = model.predict(frame_resized)
-            frame = draw_predictions(frame_resized, config['INPUT_shape'][0], pred[0], pred[1], pred[2])
+
+            frame = draw_predictions(frame_resized,
+                                     tf.cast(config['INPUT_shape'][0], tf.float32),
+                                     pred[0], pred[1], pred[2])
 
             #frame = tf.image.draw_bounding_boxes(frame[tf.newaxis, ...], XYXY_to_YXYX(pred[0][0][tf.newaxis, ...]), [[252.0, 3.0, 3.0], [18.0, 4.0, 217.0]])
